refactor(tests): use ava instead of mocha (#322)
