+++ lib/api.js
@@ -156,7 +156,7 @@ function NodeCG(bundle, socket) {
 
 	// Create read-only config property, which contains the current filtered NodeCG config
 	Object.defineProperty(this, 'config', {
-		value: filteredConfig,
+		value: JSON.parse(JSON.stringify(filteredConfig)),
 		writable: false,
 		enumerable: true
 	});
