+++ lib/replicant/replicant.js
@@ -9,8 +9,7 @@ const clone = require('clone');
 const schemaDefaults = require('json-schema-defaults');
 const replicator = require('../replicator');
 const shared = require('./shared');
-const $RefParser = require('json-schema-ref-parser');
-const sleep = require('deasync').sleep;
+const $RefParser = require('json-schema-lib');
 const REPLICANTS_ROOT = path.join(process.env.NODECG_ROOT, 'db/replicants');
 
 const REPLICANT_HANDLER = {
@@ -88,35 +87,7 @@ class Replicant extends EventEmitter {
 				path.join(process.env.NODECG_ROOT, opts.schemaPath);
 			if (fs.existsSync(absoluteSchemaPath)) {
 				try {
-					let sleepCount = 0;
-					let refParserDone = false;
-					let refParserError;
-					let refParsedSchema;
-					$RefParser.dereference(absoluteSchemaPath, {
-						dereference: {
-							circular: false
-						}
-					}, (err, schema) => {
-						refParserDone = true;
-						refParserError = err;
-						refParsedSchema = schema;
-					});
-
-					while (!refParserDone) { // eslint-disable-line no-unmodified-loop-condition
-						if (sleepCount > 150) {
-							replicator.declaredReplicants[namespace][name] = null;
-							throw new Error(`Failed to dereference schema for ${this.namespace}:${this.name} after 150 sleep cycles.`);
-						}
-
-						sleep(10);
-						sleepCount++;
-					}
-
-					if (refParserError) {
-						throw refParserError;
-					}
-
-					this.schema = refParsedSchema;
+					this.schema = $RefParser.readSync(absoluteSchemaPath).root;
 					this.schemaSum = sha1(this.schema);
 					this.validate = shared.generateValidator(this);
 				} catch (e) {
