+++ lib/server/index.js
@@ -98,44 +98,7 @@ module.exports.start = function () {
 		});
 	}
 
-	bundleManager.init(process.env.NODECG_ROOT, require('../../package.json').version, config, Logger).then(() => {
-		// Hack to restore the process title after npm has forcibly changed it
-		process.title = 'NodeCG';
-
-		log.trace('Starting dashboard lib');
-		const dashboard = require('../dashboard');
-		app.use(dashboard);
-
-		log.trace('Starting graphics lib');
-		const graphics = require('../graphics');
-		app.use(graphics);
-
-		log.trace('Starting bundle sounds lib');
-		const sounds = require('../sounds');
-		app.use(sounds);
-
-		log.trace('Starting bundle assets lib');
-		const assets = require('../assets');
-		app.use(assets);
-
-		log.trace('Starting bundle shared sources lib');
-		const sharedSources = require('../shared-sources');
-		app.use(sharedSources);
-
-		extensionManager = require('./extensions.js');
-		extensionManager.init();
-		module.exports.emit('extensionsLoaded');
-
-		// We intentionally wait until all bundles and extensions are loaded before starting the server.
-		// This has two benefits:
-		// 1) Prevents the dashboard/views from being opened before everything has finished loading
-		// 2) Prevents dashboard/views from re-declaring replicants on reconnect before extensions have had a chance
-		server.listen(config.port, config.host);
-
-		const protocol = config.ssl && config.ssl.enabled ? 'https' : 'http';
-		log.info('NodeCG running on %s://%s', protocol, config.baseURL);
-		module.exports.emit('started');
-	});
+	bundleManager.init(process.env.NODECG_ROOT, require('../../package.json').version, config, Logger);
 
 	io.on('error', err => {
 		if (global.sentryEnabled) {
@@ -259,6 +222,9 @@ module.exports.start = function () {
 	server.on('error', err => {
 		switch (err.code) {
 			case 'EADDRINUSE':
+				if (process.env.NODECG_TEST) {
+					return;
+				}
 				log.error(`[server.js] Listen ${config.host}:${config.port} in use, is NodeCG already running? NodeCG will now exit.`);
 				break;
 			default:
@@ -268,11 +234,61 @@ module.exports.start = function () {
 
 		module.exports.emit('error', err);
 	});
+
+	log.trace('Starting dashboard lib');
+	const dashboard = require('../dashboard');
+	app.use(dashboard);
+
+	log.trace('Starting graphics lib');
+	const graphics = require('../graphics');
+	app.use(graphics);
+
+	log.trace('Starting bundle sounds lib');
+	const sounds = require('../sounds');
+	app.use(sounds);
+
+	log.trace('Starting bundle assets lib');
+	const assets = require('../assets');
+	app.use(assets);
+
+	log.trace('Starting bundle shared sources lib');
+	const sharedSources = require('../shared-sources');
+	app.use(sharedSources);
+
+	extensionManager = require('./extensions.js');
+	extensionManager.init();
+	module.exports.emit('extensionsLoaded');
+
+	// We intentionally wait until all bundles and extensions are loaded before starting the server.
+	// This has two benefits:
+	// 1) Prevents the dashboard/views from being opened before everything has finished loading
+	// 2) Prevents dashboard/views from re-declaring replicants on reconnect before extensions have had a chance
+	server.listen({
+		host: config.host,
+		port: process.env.NODECG_TEST ? undefined : config.port
+	}, () => {
+		if (process.env.NODECG_TEST) {
+			const port = server.address().port;
+			log.warn(`Test mode active, using automatic listen port: ${port}`);
+			configHelper.config.port = port;
+			configHelper.filteredConfig.port = port;
+			process.env.NODECG_TEST_PORT = port;
+		}
+
+		const protocol = config.ssl && config.ssl.enabled ? 'https' : 'http';
+		log.info('NodeCG running on %s://%s', protocol, config.baseURL);
+		module.exports.emit('started');
+	});
 };
 
 module.exports.stop = function () {
-	server.close();
-	io.close();
+	if (server) {
+		server.close();
+	}
+
+	if (io) {
+		io.close();
+	}
 
 	extensionManager = null;
 	io = null;
