+++ package.json
@@ -26,7 +26,6 @@
     "clone": "^2.0.0",
     "compression": "^1.6.0",
     "cookie-parser": "^1.4.1",
-    "deasync": "^0.1.10",
     "deep-equal": "^1.0.1",
     "express": "^4.13.3",
     "express-nedb-session": "^0.1.1",
@@ -37,7 +36,7 @@
     "is-error": "^2.2.1",
     "is-my-json-valid": "^2.12.3",
     "json-schema-defaults": "^0.3.0",
-    "json-schema-ref-parser": "^3.1.2",
+    "json-schema-lib": "github:Lange/json-schema-lib#ac59fe29fad5c54647e4c483c7255a923cc430dd",
     "lodash.debounce": "^4.0.8",
     "minimist": "^1.2.0",
     "multer": "^1.1.0",
@@ -62,7 +61,7 @@
   "scripts": {
     "start": "node index.js",
     "pretest": "npm run build && npm run instrument",
-    "test": "nyc mocha -R spec && nyc report --reporter=html",
+    "test": "nyc ava && nyc report --reporter=html",
     "coveralls": "nyc report --reporter=text-lcov | coveralls",
     "static": "eslint index.js lib/**/*.js src/**/*.js test/**/*.js",
     "build": "npm run static && npm run docs:build && gulp",
@@ -92,6 +91,7 @@
   ],
   "devDependencies": {
     "aliasify": "^2.1.0",
+    "ava": "^0.22.0",
     "babel-plugin-istanbul": "^4.1.4",
     "babel-preset-babili": "^0.1.1",
     "babel-preset-es2015-nostrict": "^6.6.2",
@@ -103,6 +103,7 @@
     "del": "^2.2.2",
     "eslint": "^3.18.0",
     "eslint-config-xo": "^0.18.0",
+    "eslint-plugin-ava": "^4.2.1",
     "exorcist": "^0.4.0",
     "fs.extra": "^1.3.2",
     "gh-pages": "^1.0.0",
