+++ test/fixtures/cfg/nodecg.json
@@ -1,1 +1,6 @@
-{}
+{
+	"autodeps": {
+		"npm": false,
+		"bower": false
+	}
+}
