+++ test/helpers/nodecg-and-webdriver.js
@@ -1,16 +1,17 @@
-/* eslint-env node, mocha, browser */
-'use strict';
-
 process.env.test = true;
 process.env.NODECG_TEST = true;
+
+// Native
 const fs = require('fs');
 const path = require('path');
+
+// Packages
 const fse = require('fs-extra');
 const temp = require('temp');
-const tempFolder = temp.mkdirSync();
+const webdriverio = require('webdriverio');
 
-// Automatically track and cleanup files at exit.
-temp.track();
+const tempFolder = temp.mkdirSync();
+temp.track(); // Automatically track and cleanup files at exit.
 
 fse.copySync('test/fixtures/assets', path.join(tempFolder, 'assets'));
 fse.copySync('test/fixtures/bundles', path.join(tempFolder, 'bundles'));
@@ -21,30 +22,29 @@ fse.copySync('test/fixtures/db', path.join(tempFolder, 'db'));
 // may have. We don't want to touch any existing user data!
 process.env.NODECG_ROOT = tempFolder;
 
-const webdriverio = require('webdriverio');
-const e = require('./setup/test-environment');
-const C = require('./setup/test-constants');
-const addCustomBrowserCommands = require('./setup/custom-webdriver-commands');
+// Ours (must be after the above setup steps)
+const addCustomBrowserCommands = require('./custom-webdriver-commands');
+const C = require('./test-constants');
+const e = require('./test-environment');
 
-// Global before and after
-before(function (done) {
-	this.timeout(0);
+module.exports = function (test) {
+	test.before(async () => {
+		if (C.CONFIG.login && C.CONFIG.login.enabled) {
+			throw new Error('Login security is enabled! ' +
+				'Please disable login security in cfg/nodecg.json before running tests');
+		}
 
-	if (C.CONFIG.autodeps && (C.CONFIG.autodeps.npm === false || C.CONFIG.autodeps.bower === false)) {
-		throw new Error('Autodeps disabled! ' +
-			'Please enable auto-installing dependencies in cfg/nodecg.json before running tests');
-	}
+		if (C.CONFIG.ssl && C.CONFIG.ssl.enabled) {
+			throw new Error('SSL is enabled! Please disable SSL in cfg/nodecg.json before running tests');
+		}
 
-	if (C.CONFIG.login && C.CONFIG.login.enabled) {
-		throw new Error('Login security is enabled! ' +
-			'Please disable login security in cfg/nodecg.json before running tests');
-	}
+		process.env.NODECG_ROOT = tempFolder;
+		await new Promise((resolve, reject) => {
+			e.server.on('started', resolve);
+			e.server.on('error', reject);
+			e.server.start();
+		});
 
-	if (C.CONFIG.ssl && C.CONFIG.ssl.enabled) {
-		throw new Error('SSL is enabled! Please disable SSL in cfg/nodecg.json before running tests');
-	}
-
-	e.server.once('started', () => {
 		// Extension API setup
 		e.apis.extension = e.server.getExtensions()[C.BUNDLE_NAME];
 
@@ -116,7 +116,7 @@ before(function (done) {
 		// Attach our custom methods to this browser instance.
 		addCustomBrowserCommands(e.browser.client);
 
-		e.browser.client
+		return e.browser.client
 			.init()
 			.timeouts('script', 30000)
 			.newWindow(C.DASHBOARD_URL, 'NodeCG dashboard', '')
@@ -125,12 +125,16 @@ before(function (done) {
 				e.browser.tabs.dashboard = tabId;
 			})
 			.executeAsync(done => {
+				/* eslint-disable no-undef */
 				const checkForApi = setInterval(() => {
-					if (typeof window.dashboardApi !== 'undefined') {
+					if (typeof window.dashboardApi !== 'undefined' && typeof Polymer !== 'undefined') {
 						clearInterval(checkForApi);
-						done();
+						Polymer.RenderStatus.afterNextRender(this, () => {
+							done();
+						});
 					}
 				}, 50);
+				/* eslint-enable no-undef */
 			})
 			.newWindow(`${C.TEST_PANEL_URL}?standalone=true`, 'NodeCG test bundle standalone panel', '')
 			.getCurrentTabId()
@@ -158,41 +162,40 @@ before(function (done) {
 					}
 				}, 50);
 			})
-			.timeouts('script', 5000)
-			.call(done)
-			.catch(done);
+			.timeouts('script', 5000);
 	});
 
-	e.server.start();
-});
-
-after(async function () {
-	this.timeout(15000);
-
-	/* eslint-disable no-await-in-loop */
-	for (const tabName in e.browser.tabs) {
-		if (!{}.hasOwnProperty.call(e.browser.tabs, tabName)) {
-			continue;
+	test.after.always(async () => {
+		if (e.server) {
+			e.server.stop();
 		}
 
-		await e.browser.client.switchTab(e.browser.tabs[tabName]);
-		const {value: coverageObj} = await e.browser.client.execute('return window.__coverage__;');
+		if (e.browser && e.browser.client && typeof e.browser.client.end === 'function') {
+			/* eslint-disable no-await-in-loop */
+			for (const tabName in e.browser.tabs) {
+				if (!{}.hasOwnProperty.call(e.browser.tabs, tabName)) {
+					continue;
+				}
 
-		const newCoverageObj = {};
-		for (const key in coverageObj) {
-			if (!{}.hasOwnProperty.call(e.browser.tabs, tabName)) {
-				continue;
+				await e.browser.client.switchTab(e.browser.tabs[tabName]);
+				const {value: coverageObj} = await e.browser.client.execute('return window.__coverage__;');
+
+				const newCoverageObj = {};
+				for (const key in coverageObj) {
+					if (!{}.hasOwnProperty.call(e.browser.tabs, tabName)) {
+						continue;
+					}
+
+					const absKey = path.resolve('src', key);
+					coverageObj[key].path = absKey;
+					newCoverageObj[absKey] = coverageObj[key];
+				}
+
+				fs.writeFileSync(`.nyc_output/browser-${tabName}.json`, JSON.stringify(newCoverageObj), 'utf-8');
 			}
+			/* eslint-enable no-await-in-loop */
 
-			const absKey = path.resolve('src', key);
-			coverageObj[key].path = absKey;
-			newCoverageObj[absKey] = coverageObj[key];
+			return e.browser.client.end();
 		}
-
-		fs.writeFileSync(`.nyc_output/browser-${tabName}.json`, JSON.stringify(newCoverageObj), 'utf-8');
-	}
-	/* eslint-enable no-await-in-loop */
-
-	e.server.stop();
-	return e.browser.client.end();
-});
+	});
+};
